Commit4

+++ app/src/main/java/es/iessaladillo/pedrojoya/baldogym/ui/trainingsession/TrainingSessionActivityViewModelFactory.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.ViewModelProvider
 import es.iessaladillo.pedrojoya.baldogym.data.LocalRepository
 import es.iessaladillo.pedrojoya.baldogym.ui.schedule.ScheduleActivityViewModel
 
-class TrainingSessionActivityViewModelFactory(private val localRepository: LocalRepository, val application: Application, private val id:Int) : ViewModelProvider.Factory{
+class TrainingSessionActivityViewModelFactory(private val localRepository: LocalRepository, val application: Application, private val id:Long) : ViewModelProvider.Factory{
 
     override fun <T : ViewModel?> create(modelClass: Class<T>): T {
         return TrainingSessionActivityViewModel(localRepository,application,id) as T
